implemented resolveImplementedSuggestionsOnPlatform

+++ libs/code-review/pipeline/stages/process-files-review.stage.ts
@@ -34,8 +34,6 @@ import {
     IFinalAnalysisResult,
 } from '@libs/core/infrastructure/config/types/general/codeReview.type';
 import { createOptimizedBatches } from '@libs/common/utils/batch.helper';
-import { DeliveryStatus } from '@libs/platformData/domain/pullRequests/enums/deliveryStatus.enum';
-import { ImplementationStatus } from '@libs/platformData/domain/pullRequests/enums/implementationStatus.enum';
 import { PriorityStatus } from '@libs/platformData/domain/pullRequests/enums/priorityStatus.enum';
 import { TaskStatus } from '@libs/ee/kodyAST/interfaces/code-ast-analysis.interface';
 import { OrganizationAndTeamData } from '@libs/core/infrastructure/config/types/general/organizationAndTeamData';
@@ -673,7 +671,6 @@ export class ProcessFilesReview extends BasePipelineStage<CodeReviewPipelineCont
         ];
 
         // If it's a commit, validate repeated suggestions
-        // NOTE: This logic is also handled asynchronously by ImplementationVerificationProcessor
         // We keep it here to ensure immediate feedback within the review session context, but avoid race conditions
         // by checking if the async job will likely cover it (though redundancy is generally safe as operations are idempotent).
         if (context?.action && VALID_ACTIONS.includes(context.action)) {
@@ -685,33 +682,14 @@ export class ProcessFilesReview extends BasePipelineStage<CodeReviewPipelineCont
                     context.organizationAndTeamData,
                 );
 
-            if (savedSuggestions?.length > 0) {
-                const sentSuggestions = savedSuggestions.filter(
-                    (suggestion) =>
-                        suggestion.deliveryStatus === DeliveryStatus.SENT &&
-                        suggestion.implementationStatus ===
-                            ImplementationStatus.NOT_IMPLEMENTED,
-                );
-
-                if (mergedSuggestions?.length > 0) {
-                    mergedSuggestions =
-                        await this.suggestionService.removeSuggestionsRelatedToSavedFiles(
-                            context?.organizationAndTeamData,
-                            context?.pullRequest?.number.toString(),
-                            savedSuggestions,
-                            mergedSuggestions,
-                        );
-                }
-
-                // We can only validate the implementation of suggestions that were sent
-                if (sentSuggestions.length > 0) {
-                    await this.suggestionService.validateImplementedSuggestions(
+            if (savedSuggestions?.length > 0 && mergedSuggestions?.length > 0) {
+                mergedSuggestions =
+                    await this.suggestionService.removeSuggestionsRelatedToSavedFiles(
                         context?.organizationAndTeamData,
-                        file?.patch,
-                        sentSuggestions,
-                        context?.pullRequest?.number,
+                        context?.pullRequest?.number.toString(),
+                        savedSuggestions,
+                        mergedSuggestions,
                     );
-                }
             }
         }
